requirements.txt
# Polymarket Sports Trading Bot - Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
aiosqlite>=0.19.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# HTTP Client
httpx>=0.26.0

# WebSocket
websocket-client>=1.7.0

# Polymarket SDK
py-clob-client>=0.1.0

# Templates
jinja2>=3.1.0

# Task Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
